kaikki stepdefit samaan tiedostoon
